+++ client/components/Host/thumbnail.jsx
